moves into auth folder
